extract item-rows component
