+++ apps/frontend/src/components/forms/newDonationFormModal.tsx
@@ -1,36 +1,14 @@
-import {
-  Button,
-  Text,
-  Flex,
-  Box,
-  Table,
-  Input,
-  TableCaption,
-  Stack,
-  Dialog,
-  NativeSelect,
-  NativeSelectIndicator,
-  Portal,
-  Checkbox,
-  Menu,
-  NumberInput,
-  Tooltip,
-} from '@chakra-ui/react';
-import { useState } from 'react';
+import { Text, Dialog } from '@chakra-ui/react';
 import ApiClient from '@api/apiClient';
-import {
-  CreateDonationDto,
-  DayOfWeek,
-  FoodType,
-  RecurrenceEnum,
-  RepeatOnState,
-} from '../../types/types';
-import { Minus } from 'lucide-react';
-import { generateNextDonationDate } from '@utils/utils';
+import { CreateDonationDto, FoodType, RecurrenceEnum } from '../../types/types';
 import { FloatingAlert } from '@components/floatingAlert';
 import { useAlert } from '../../hooks/alert';
 import { useModalBodyCleanup } from '../../hooks/modalBodyCleanup';
 import { AlertStatus } from '../../types/types';
+import EditableDonationItemsTable, {
+  DonationRow,
+  RecurrenceData,
+} from './editableDonationItemsTable';
 
 interface NewDonationFormModalProps {
   onDonationSuccess: () => void;
@@ -38,192 +16,24 @@ interface NewDonationFormModalProps {
   onClose: () => void;
 }
 
-interface DonationRow {
-  id: number;
-  foodItem: string;
-  foodType: FoodType | '';
-  numItems: string;
-  ozPerItem: string;
-  valuePerItem: string;
-  foodRescue: boolean;
-}
-
-// Display labels for RecurrenceEnum values in the UI
-const RECURRENCE_LABELS: Record<RecurrenceEnum, string> = {
-  [RecurrenceEnum.NONE]: 'None',
-  [RecurrenceEnum.WEEKLY]: 'Week',
-  [RecurrenceEnum.MONTHLY]: 'Month',
-  [RecurrenceEnum.YEARLY]: 'Year',
-};
-
-// Ensure valid decimals and positive integers for input validation
-const isValidDecimal = (val: string): boolean =>
-  val !== '' && /^\d+(\.\d{1,2})?$/.test(val) && parseFloat(val) > 0;
-const isValidPositiveInt = (val: string): boolean =>
-  val !== '' && /^\d+$/.test(val) && parseInt(val) > 0;
-
-// Displays appropriate tooltip if necessary
-const getFirstValidationError = (
-  rows: DonationRow[],
-  isRecurring: boolean,
-  repeatEvery: string,
-  endsAfter: string,
-): string | null => {
-  for (let i = 0; i < rows.length; i++) {
-    const row = rows[i];
-    const rowLabel = rows.length > 1 ? ` (row ${i + 1})` : '';
-
-    if (row.foodItem.trim() === '') {
-      return `Food item${rowLabel} is required.`;
-    }
-    if (row.foodType === '') {
-      return `Food type${rowLabel} is required.`;
-    }
-    if (row.numItems === '') {
-      return `Quantity${rowLabel} is required.`;
-    }
-    if (!isValidPositiveInt(row.numItems)) {
-      return `Quantity${rowLabel} must be a positive whole number.`;
-    }
-    if (row.ozPerItem === '') {
-      return `Oz. per item${rowLabel} is required.`;
-    }
-    if (!isValidDecimal(row.ozPerItem)) {
-      return `Oz. per item${rowLabel} must be a positive number with at most 2 decimal places.`;
-    }
-    if (row.valuePerItem === '') {
-      return `Donation value${rowLabel} is required.`;
-    }
-    if (!isValidDecimal(row.valuePerItem)) {
-      return `Donation value${rowLabel} must be a positive number with at most 2 decimal places.`;
-    }
-  }
-  if (isRecurring) {
-    if (!isValidPositiveInt(repeatEvery)) {
-      return 'Repeat every must be a positive whole number.';
-    }
-    if (!isValidPositiveInt(endsAfter)) {
-      return 'Ends after must be a positive whole number.';
-    }
-  }
-  return null;
-};
-
 const NewDonationFormModal: React.FC<NewDonationFormModalProps> = ({
   onDonationSuccess,
   isOpen,
   onClose,
 }) => {
   useModalBodyCleanup();
-  const [rows, setRows] = useState<DonationRow[]>([
-    {
-      id: 1,
-      foodItem: '',
-      foodType: '',
-      numItems: '',
-      ozPerItem: '',
-      valuePerItem: '',
-      foodRescue: false,
-    },
-  ]);
 
-  const [isRecurring, setIsRecurring] = useState(false);
-  const [repeatEvery, setRepeatEvery] = useState('1');
-  const [repeatInterval, setRepeatInterval] = useState<RecurrenceEnum>(
-    RecurrenceEnum.NONE,
-  );
-  const [repeatOn, setRepeatOn] = useState<RepeatOnState>({
-    Monday: false,
-    Tuesday: false,
-    Wednesday: true,
-    Thursday: false,
-    Friday: false,
-    Saturday: false,
-    Sunday: false,
-  });
-  const [endsAfter, setEndsAfter] = useState('1');
   const [alertState, setAlertMessage] = useAlert();
 
-  const handleChange = (id: number, field: string, value: string | boolean) => {
-    const updatedRows = rows.map((row) =>
-      row.id === id ? { ...row, [field]: value } : row,
-    );
-    setRows(updatedRows);
-  };
-
-  const handleDayToggle = (day: DayOfWeek) => {
-    setRepeatOn((prev) => ({ ...prev, [day]: !prev[day] }));
-  };
-
-  const addRow = () => {
-    setRows([
-      ...rows,
-      {
-        id: Date.now(),
-        foodItem: '',
-        foodType: '',
-        numItems: '',
-        ozPerItem: '',
-        valuePerItem: '',
-        foodRescue: false,
-      },
-    ]);
-  };
-
-  const deleteRow = (id: number) => {
-    if (rows.length > 1) {
-      const newRows = rows.filter((r) => r.id !== id);
-      setRows(newRows);
-    }
-  };
-
-  const getNextDonationDateDisplay = (): string => {
-    const date = generateNextDonationDate(
-      repeatEvery,
-      repeatInterval,
-      repeatOn,
-    );
-    if (!date) return '';
-    return new Date(date).toLocaleDateString('en-US', {
-      weekday: 'long',
-      year: 'numeric',
-      month: 'long',
-      day: 'numeric',
-    });
-  };
-
-  const getSelectedDaysText = () => {
-    const selected = (Object.keys(repeatOn) as DayOfWeek[]).filter(
-      (day) => repeatOn[day],
-    );
-    if (selected.length === 0) return 'Select days';
-    if (selected.length === 1) return selected[0];
-    if (selected.length <= 4) return selected.join(', ');
-    return `${selected.slice(0, 4).join(', ')} + ${selected.length - 4}`;
-  };
-
-  const validateAndSubmit = async () => {
-    // Recurring: weekly day selection
-    if (
-      isRecurring &&
-      repeatInterval === RecurrenceEnum.WEEKLY &&
-      !Object.values(repeatOn).some(Boolean)
-    ) {
-      setAlertMessage(
-        'Please select at least one day for weekly recurrence.',
-        AlertStatus.ERROR,
-      );
-      return;
-    }
-
+  const handleLogDonation = async (
+    rows: DonationRow[],
+    recurrence: RecurrenceData | null,
+  ) => {
     const donationBody: CreateDonationDto = {
-      recurrenceFreq: isRecurring ? parseInt(repeatEvery) : undefined,
-      recurrence: isRecurring ? repeatInterval : RecurrenceEnum.NONE,
-      repeatOnDays:
-        isRecurring && repeatInterval === RecurrenceEnum.WEEKLY
-          ? repeatOn
-          : undefined,
-      occurrencesRemaining: isRecurring ? parseInt(endsAfter) : undefined,
+      recurrenceFreq: recurrence?.recurrenceFreq,
+      recurrence: recurrence?.recurrence ?? RecurrenceEnum.NONE,
+      repeatOnDays: recurrence?.repeatOnDays,
+      occurrencesRemaining: recurrence?.occurrencesRemaining,
       items: rows.map((row) => ({
         itemName: row.foodItem,
         quantity: parseInt(row.numItems),
@@ -237,40 +47,10 @@ const NewDonationFormModal: React.FC<NewDonationFormModalProps> = ({
     try {
       await ApiClient.postDonation(donationBody);
       onDonationSuccess();
-
-      setRows([
-        {
-          id: 1,
-          foodItem: '',
-          foodType: '',
-          numItems: '',
-          ozPerItem: '',
-          valuePerItem: '',
-          foodRescue: false,
-        },
-      ]);
-      setIsRecurring(false);
-      setRepeatInterval(RecurrenceEnum.NONE);
       onClose();
     } catch {
       setAlertMessage('Error submitting new donation', AlertStatus.ERROR);
     }
-  };
-
-  const firstValidationError = getFirstValidationError(
-    rows,
-    isRecurring,
-    repeatEvery,
-    endsAfter,
-  );
-  const isSubmitDisabled = firstValidationError !== null;
-  const isRepeatOnDisabled = repeatInterval !== RecurrenceEnum.WEEKLY;
-
-  const placeholderStyles = {
-    color: 'neutral.300',
-    fontFamily: 'inter',
-    fontSize: 'sm',
-    fontWeight: '400',
   };
 
   return (
@@ -311,444 +91,12 @@ const NewDonationFormModal: React.FC<NewDonationFormModalProps> = ({
               Value.
             </Text>
 
-            <Box display="block" overflowX="auto" whiteSpace="nowrap">
-              <Table.Root
-                variant="line"
-                size="md"
-                style={{ borderCollapse: 'collapse' }}
-              >
-                <TableCaption textAlign="left">
-                  <Stack direction="column" align="flex-start" gap={14} mt={4}>
-                    <Button
-                      display="inline-flex"
-                      alignItems="center"
-                      bg="white"
-                      color="neutral.800"
-                      fontWeight={600}
-                      fontSize={14}
-                      borderRadius={4}
-                      borderColor="neutral.200"
-                      onClick={addRow}
-                    >
-                      Add New Row +
-                    </Button>
-                    <Checkbox.Root
-                      checked={isRecurring}
-                      onCheckedChange={(e: { checked: boolean }) => {
-                        setIsRecurring(!!e.checked);
-                        setRepeatInterval(
-                          e.checked
-                            ? RecurrenceEnum.WEEKLY
-                            : RecurrenceEnum.NONE,
-                        );
-                      }}
-                    >
-                      <Checkbox.HiddenInput />
-                      <Checkbox.Control>
-                        <Checkbox.Indicator />
-                      </Checkbox.Control>
-                      <Checkbox.Label color="neutral.700" fontWeight={400}>
-                        Make Donation Recurring Reminders
-                      </Checkbox.Label>
-                    </Checkbox.Root>
-                  </Stack>
-                </TableCaption>
-
-                <Table.Header>
-                  <Table.Row fontWeight={600}>
-                    <Table.ColumnHeader width="32px" p={1} />
-                    <Table.ColumnHeader width="23%" p={0}>
-                      Food Item
-                      <Text as="span" color="red">
-                        *
-                      </Text>
-                    </Table.ColumnHeader>
-                    <Table.ColumnHeader width="22%">
-                      Food Type
-                      <Text as="span" color="red">
-                        *
-                      </Text>
-                    </Table.ColumnHeader>
-                    <Table.ColumnHeader width="14%">
-                      Quantity
-                      <Text as="span" color="red">
-                        *
-                      </Text>
-                    </Table.ColumnHeader>
-                    <Table.ColumnHeader width="14%">
-                      Oz. per item
-                      <Text as="span" color="red">
-                        *
-                      </Text>
-                    </Table.ColumnHeader>
-                    <Table.ColumnHeader width="14%">
-                      <Text>Donation Value</Text>
-                      (Fair Market Value of Food Only)
-                      <Text as="span" color="red">
-                        *
-                      </Text>
-                    </Table.ColumnHeader>
-                    <Table.ColumnHeader
-                      width="20px"
-                      textAlign="left"
-                      px={0}
-                      pl={4}
-                      whiteSpace="normal"
-                      lineHeight="tight"
-                    >
-                      Food Rescue
-                      <Text as="span" color="red">
-                        *
-                      </Text>
-                    </Table.ColumnHeader>
-                  </Table.Row>
-                </Table.Header>
-
-                <Table.Body>
-                  {rows.map((row) => (
-                    <Table.Row key={row.id}>
-                      <Table.Cell width="32px" p={0} pr={2}>
-                        <Button
-                          size="sm"
-                          variant="outline"
-                          onClick={() => deleteRow(row.id)}
-                          disabled={rows.length === 1}
-                          borderColor="neutral.300"
-                          borderRadius="md"
-                          bg="white"
-                          _hover={{ bg: 'neutral.50' }}
-                          _disabled={{ opacity: 0.4, cursor: 'not-allowed' }}
-                          width="28px"
-                          height="28px"
-                          minW="28px"
-                          padding={0}
-                        >
-                          <Box color="neutral.300">
-                            <Minus style={{ width: '24px', height: '24px' }} />
-                          </Box>
-                        </Button>
-                      </Table.Cell>
-
-                      <Table.Cell p={0} pr={4}>
-                        <Input
-                          _placeholder={placeholderStyles}
-                          color="neutral.800"
-                          placeholder="Enter Food"
-                          value={row.foodItem}
-                          onChange={(e) =>
-                            handleChange(row.id, 'foodItem', e.target.value)
-                          }
-                        />
-                      </Table.Cell>
-
-                      <Table.Cell>
-                        <NativeSelect.Root size="md" width="100%">
-                          <NativeSelect.Field
-                            color={row.foodType ? 'neutral.800' : 'neutral.300'}
-                            placeholder="Select Type"
-                            value={row.foodType}
-                            onChange={(e) =>
-                              handleChange(row.id, 'foodType', e.target.value)
-                            }
-                          >
-                            {Object.values(FoodType).map((type) => (
-                              <option
-                                key={type}
-                                value={type}
-                                style={{
-                                  color: 'var(--chakra-colors-neutral-800)',
-                                }}
-                              >
-                                {type}
-                              </option>
-                            ))}
-                          </NativeSelect.Field>
-                          <NativeSelectIndicator />
-                        </NativeSelect.Root>
-                      </Table.Cell>
-
-                      <Table.Cell>
-                        <Input
-                          _placeholder={placeholderStyles}
-                          color="neutral.800"
-                          placeholder="Enter #"
-                          type="number"
-                          min={1}
-                          step={1}
-                          value={row.numItems}
-                          onChange={(e) =>
-                            handleChange(row.id, 'numItems', e.target.value)
-                          }
-                        />
-                      </Table.Cell>
-
-                      <Table.Cell>
-                        <Input
-                          _placeholder={placeholderStyles}
-                          color="neutral.800"
-                          placeholder="Enter #"
-                          type="number"
-                          min={0.01}
-                          step={0.01}
-                          value={row.ozPerItem}
-                          onChange={(e) =>
-                            handleChange(row.id, 'ozPerItem', e.target.value)
-                          }
-                        />
-                      </Table.Cell>
-
-                      <Table.Cell>
-                        <Input
-                          _placeholder={placeholderStyles}
-                          color="neutral.800"
-                          placeholder="Enter $"
-                          type="number"
-                          min={0.01}
-                          step={0.01}
-                          value={row.valuePerItem}
-                          onChange={(e) =>
-                            handleChange(row.id, 'valuePerItem', e.target.value)
-                          }
-                        />
-                      </Table.Cell>
-
-                      <Table.Cell px={0} pl={6} width="32px">
-                        <Checkbox.Root
-                          checked={row.foodRescue}
-                          size="lg"
-                          borderRadius="2px"
-                          onCheckedChange={(e: { checked: boolean }) =>
-                            handleChange(row.id, 'foodRescue', !!e.checked)
-                          }
-                        >
-                          <Checkbox.HiddenInput />
-                          <Checkbox.Control
-                            borderRadius="2px"
-                            borderColor="#E4E4E7"
-                          >
-                            <Checkbox.Indicator />
-                          </Checkbox.Control>
-                        </Checkbox.Root>
-                      </Table.Cell>
-                    </Table.Row>
-                  ))}
-                </Table.Body>
-              </Table.Root>
-            </Box>
-
-            {isRecurring && (
-              <Box mt={6} color="neutral.800" fontSize="sm">
-                <Stack
-                  direction="row"
-                  align="flex-start"
-                  gap={4}
-                  mb={4}
-                  flexWrap="wrap"
-                >
-                  <Box flex="1" minW="200px">
-                    <Text fontWeight={600} mb={3}>
-                      Repeat every
-                    </Text>
-                    <Flex gap={2} align="center">
-                      <NumberInput.Root
-                        width="98px"
-                        value={repeatEvery}
-                        onValueChange={(e: { value: string }) =>
-                          setRepeatEvery(e.value)
-                        }
-                        min={1}
-                        step={1}
-                      >
-                        <NumberInput.Input />
-                        <NumberInput.Control />
-                      </NumberInput.Root>
-                      <NativeSelect.Root flex="1" size="md">
-                        <NativeSelect.Field
-                          value={repeatInterval}
-                          onChange={(e) =>
-                            setRepeatInterval(e.target.value as RecurrenceEnum)
-                          }
-                        >
-                          {(Object.values(RecurrenceEnum) as RecurrenceEnum[])
-                            .filter((v) => v !== RecurrenceEnum.NONE)
-                            .map((v) =>
-                              repeatEvery === '1' ? (
-                                <option key={v} value={v}>
-                                  {RECURRENCE_LABELS[v]}
-                                </option>
-                              ) : (
-                                <option key={v} value={v}>
-                                  {RECURRENCE_LABELS[v]}s
-                                </option>
-                              ),
-                            )}
-                        </NativeSelect.Field>
-                        <NativeSelectIndicator />
-                      </NativeSelect.Root>
-                    </Flex>
-                  </Box>
-
-                  <Box flex="1" minW="236px" width="100%">
-                    <Text fontWeight={600} mb={3}>
-                      Repeat on
-                    </Text>
-                    <Menu.Root
-                      closeOnSelect={false}
-                      positioning={{ sameWidth: true }}
-                    >
-                      {!isRepeatOnDisabled ? (
-                        <Menu.Trigger asChild>
-                          <Box cursor="pointer" width="100%">
-                            <NativeSelect.Root size="md" width="100%">
-                              <NativeSelect.Field
-                                value={getSelectedDaysText()}
-                                bg="white"
-                                readOnly
-                              >
-                                <option>{getSelectedDaysText()}</option>
-                              </NativeSelect.Field>
-                              <NativeSelectIndicator />
-                            </NativeSelect.Root>
-                          </Box>
-                        </Menu.Trigger>
-                      ) : (
-                        <Box cursor="not-allowed">
-                          <NativeSelect.Root size="md">
-                            <NativeSelect.Field
-                              value={getSelectedDaysText()}
-                              bg="white"
-                              disabled
-                              opacity={0.5}
-                              readOnly
-                              pointerEvents="none"
-                            >
-                              <option>{getSelectedDaysText()}</option>
-                            </NativeSelect.Field>
-                            <NativeSelectIndicator />
-                          </NativeSelect.Root>
-                        </Box>
-                      )}
-                      {!isRepeatOnDisabled && (
-                        <Portal>
-                          <Menu.Positioner>
-                            <Menu.Content
-                              maxH="300px"
-                              color="neutral.800"
-                              zIndex={9999}
-                            >
-                              {(Object.keys(repeatOn) as DayOfWeek[]).map(
-                                (day) => (
-                                  <Menu.Item
-                                    key={day}
-                                    value={day}
-                                    onClick={() => handleDayToggle(day)}
-                                    p={2}
-                                  >
-                                    <Flex align="center" gap={2}>
-                                      <Checkbox.Root
-                                        checked={repeatOn[day]}
-                                        pointerEvents="none"
-                                      >
-                                        <Checkbox.HiddenInput />
-                                        <Checkbox.Control>
-                                          <Checkbox.Indicator />
-                                        </Checkbox.Control>
-                                      </Checkbox.Root>
-                                      <Text>{day}</Text>
-                                    </Flex>
-                                  </Menu.Item>
-                                ),
-                              )}
-                            </Menu.Content>
-                          </Menu.Positioner>
-                        </Portal>
-                      )}
-                    </Menu.Root>
-                  </Box>
-
-                  <Box flex="1" minW="200px">
-                    <Text fontWeight={600} mb={3} fontSize="sm">
-                      Ends after
-                    </Text>
-                    <NumberInput.Root
-                      width="100%"
-                      value={endsAfter}
-                      onValueChange={(e: { value: string }) =>
-                        setEndsAfter(e.value)
-                      }
-                      min={1}
-                      step={1}
-                    >
-                      <Flex position="relative" align="center">
-                        <NumberInput.Input pl={4} pr="140px" fontSize="sm" />
-                        <Text
-                          position="absolute"
-                          left={`calc(16px + ${
-                            endsAfter.length * 0.6
-                          }em + 8px)`}
-                          color="neutral.800"
-                          fontSize="sm"
-                          pointerEvents="none"
-                        >
-                          {parseInt(endsAfter) > 1 ? 'Reminders' : 'Reminder'}
-                        </Text>
-                        <NumberInput.Control />
-                      </Flex>
-                    </NumberInput.Root>
-                  </Box>
-                </Stack>
-
-                {(repeatInterval !== RecurrenceEnum.WEEKLY ||
-                  Object.values(repeatOn).some(Boolean)) && (
-                  <Text color="neutral.700" fontStyle="italic" mt={2}>
-                    Next donation reminder scheduled for{' '}
-                    {getNextDonationDateDisplay()}
-                  </Text>
-                )}
-              </Box>
-            )}
-
-            <Flex
-              justifyContent="flex-end"
-              gap={3}
-              mt={6}
-              pt={4}
-              align="center"
-            >
-              <Button
-                variant="outline"
-                color="gray.700"
-                fontWeight={600}
-                onClick={onClose}
-                size="md"
-              >
-                Cancel
-              </Button>
-              <Tooltip.Root disabled={!isSubmitDisabled} openDelay={0}>
-                <Tooltip.Trigger asChild>
-                  <Box display="inline-block">
-                    <Button
-                      backgroundColor="blue.ssf"
-                      onClick={validateAndSubmit}
-                      size="md"
-                      fontWeight={600}
-                      disabled={isSubmitDisabled}
-                      _disabled={{ opacity: 0.4, cursor: 'not-allowed' }}
-                      pointerEvents={isSubmitDisabled ? 'none' : 'auto'}
-                    >
-                      Submit Donation
-                    </Button>
-                  </Box>
-                </Tooltip.Trigger>
-                <Portal>
-                  <Tooltip.Positioner>
-                    <Tooltip.Content>
-                      {firstValidationError ?? ''}
-                    </Tooltip.Content>
-                  </Tooltip.Positioner>
-                </Portal>
-              </Tooltip.Root>
-            </Flex>
+            <EditableDonationItemsTable
+              showRecurrence
+              onCancel={onClose}
+              onSubmit={handleLogDonation}
+              submitButtonLabel="Submit Donation"
+            ></EditableDonationItemsTable>
           </Dialog.Body>
         </Dialog.Content>
       </Dialog.Positioner>
